+++ template/vitest.config.ts
@@ -1,0 +1,7 @@
+// <reference types="vitest" />
+/// <reference types="vite/client" />
+
+import { defineConfig } from "vitest/config";
+
+export default defineConfig({
+});